# Conflicts:
#	gns3server/crash_report.py
#	gns3server/static/web-ui/index.html
#	gns3server/version.py
#	tests/compute/docker/test_docker_vm.py

+++ gns3server/controller/__init__.py
@@ -317,12 +317,12 @@ class Controller:
         else:
             for entry in importlib_resources.files('gns3server').joinpath(resource_name).iterdir():
                 full_path = os.path.join(dst_path, entry.name)
-                if not os.path.exists(full_path):
-                    if entry.is_file():
-                        log.debug(f'Installing {resource_name} resource file "{entry.name}" to "{full_path}"')
-                        shutil.copy(str(entry), os.path.join(dst_path, entry.name))
-                    elif entry.is_dir():
-                        os.makedirs(full_path, exist_ok=True)
+                if entry.is_file() and not os.path.exists(full_path):
+                    log.debug(f'Installing {resource_name} resource file "{entry.name}" to "{full_path}"')
+                    shutil.copy(str(entry), os.path.join(dst_path, entry.name))
+                elif entry.is_dir():
+                    os.makedirs(full_path, exist_ok=True)
+                    Controller.install_resource_files(full_path, os.path.join(resource_name, entry.name))
 
     def _install_base_configs(self):
         """
